Added run and debug targets to CMake and a Makefile for convenience
The run target just runs QEMU with the kernel file from the sysroot.
The debug target additionally opens the remote connection in QEMU and
hooks a gdb instance into it for remote debugging.
The utility Makefile just makes it easier to run CMake and the created
Make targets from the project root.