Implemented a basic bitmap-based heap and the corresponding interface
The linker now also includes space for a .heap section in the .bss section.
This .heap section corresponds with a 2MiB space defined in kmalloc.cpp
reserved for the initial heap. kmalloc.hpp serves as an interface to the
underlying heap implementation currently consisting of only a
bitmap-based heap. It also implements the C++ standard new and delete
operators. This makes it possible to extend the heap implementation with
an expandable heap when we have a memory manager, or to implement a SLAB
allocator for fast small allocations.

In the entry.cpp there is a basic test for the heap that needs to be
expanded.

TODO: Seperate tests from entry.cpp