Added IO port interface and i686 implementation
The IO interface defines basic methods for 1-byte, 2-byte and 4-byte in
and out instructions and templated functions for convenience. It also
defines a class that represents an IO port. Due to the current code
structure inlining these methods is not possible.

TODO: Find method to inline such architecture specific function
      (maybe with architecture specific include guards?)