Implemented VGA Text Mode interface and corresponding kstdio methods
The VGA Text Mode interface consists of puts and putc methods and a way
to change the text color. This is used in the kstdio interface to print
messages. The kstdio interface will be expanded to more interfaces like
Serial communication. Systems like the heap can now print linux-like
init messages.