Restructured some files and applied new .clang-format
Architecture specific code for IO-opcodes where moved to an i686 folder.
This is now included with include-guards in a general io.hpp header.

More consistency with the use of `extern "C"` which now is used with C
library functions and where it is necessary.

Moved kmalloc.hpp from heap to libk.

Cleaned up more code and the CMakeLists.txt files.