Implemented interrupts and a basic interface for registering them
Interrupts are implemented so that the interrupts number and various
general registers are available for the interrupt handler. To enable
this, each interrupt needed a unique entry point that pushes the
interrupt number onto the stack. These methods are defined with macros
in the interrupt_stubs.hpp file. All these stubs call one common
interrupt entry point that pushes all generic registers and sets up the
kernel code and data segment registers. The common interrupt handler
method now executes all registered functions and acknowledges the
interrupt (e.g. EOI for PIC). If no action is registered for an
interrupt an unhandled interrupt kernel panic is thrown.