Implemented Serial communication
Kernel messages now also try to print to the COM1 Serial port. This is
used with QEMU which can redirect Serial data to stdio. There are only
four instances of the SerialPort class as there are only four Serial
ports (COM1 - COM4). By default Serial ports are configured with a Baud
rate of 115200Hz a data length of 8, one stop bit and no parity bits.