Implememted a simle PCI interface with enumeration of PCI busses
The Function class defines a function of a device, it provides functions
to access common values of its Configuration Space. The HostBridge class
is a Singleton implementation of a device function and respresents the
Host Bridge Adapter at PCI address 00:00.0. It currently only has
functions for enumerating all devices present on the PCI bus. It uses a
recursive method to find all PCI busses by determining PCI-to-PCI
Bridges and enumerating them. All connected devices are stored in a
Vector in the HostBridge class.