Worked on the C++ Library some more and renamed Vector and String
Implemented select parts of the type_traits header and utility header
needed for the functional header implementation.

Wrote a rudimentary functional header implementation.

Implemented parts of the iterator header and wrote a simple
random-access iterator implementation used for example for the vector
class.

Made the Vector and String class lowercase in conformance with the C++
Standard.