Added VSCode specific settings and launch configurations
The settings define basic stuff for the C/C++ extenstion and default
defines for the architecture and debug statements.

The files contain a launch configuration to debug the kernel inside
VSCode, it attaches to a running QEMU client, which is why a script to
run QEMU is provided. The script should be run before debugging.