Fixed smaller bugs in LibK and added some more functions
Fixed wrong signedness in the vector class.
Added more rounding functions to kmath.hpp.
Added more constructors to the function class. This is still not perfect
because the compiler still won't compile function instances passed by
value.
Added a typedef for nullptr_t in kcstddef.hpp as we cannot use the
freestanding libstdc++ which relies on some runtime functions while
cross-compiling.
Added a function to get the number of elements in an array at compile time.