Implemented katomic.hpp to be able to use atomic objects
Implemented the atomic operations using the GCC __atomic_xxx built-ins.
Implemented an atomic struct based on the C++ reference and made the
memory orders available with the MemoryOrder enum.

The atomic struct still needs to be implemented for pointer types.

These will be used when we implement Spinlocks, which will in turn make
it possible to implement Mutexes and Semaphores later.