Added a simple Spinlock
The Spinlock is implemented using an atomic_int which will store 1 if
the Spinlock is locked and 0 otherwise. On lock() interrupts are
disabled and on unlock() they will be reenabled.

For x86 architectures the lock object is aligned on an 64-byte boundary
to keep each Spinlock on a different cache-line.