Implemented a basic SMP messaging system
It utilizes IPIs to notify cores of new SMP messages. SMP messages are
queued up on the core, and an IPI to that core can be issued with
smp_poke(). For this the IPI interrupt handler was moved into the
Processor code and changes had to be made to the initialization code.
Furthermore InterruptHandler now holds a bitmap of the cores it is
registered on.