Added some improvements for the SMP system
Firstly, there was a bug in the stack tracing code, where the stack
would be misaligned betweeen faults and normal interrupts. This is
probably because I messed up somewhere in the isr stubs with aligning
the stack layouts.
The debug print code now prints the core it is currently running on.
The page fault handler now prints the address of the instruction that
caused it.