Userspace threads using TSS, Syscalls and a VMM overhaul
This commit implements userspace threads and syscalls. Necessary for
that was an overhaul to the Virtual Memory Manager. It now has a more
concise API where one can either allocate regions, where the physical
memory gets allocated using the Physical Memory Manager. Or one can
directly map physical to virtual memory. Along with that there are some
helper methods to do identity mapping or to map something based on the
size of its type. Each method takes a mapping_config_t which determines
if the regions should be user accesible, cacheable, etc. Multiple paging
spaces are now also supported which is needed as each userspace process
will need to have its own memory space.

Userspace threads are now supported. They utilise the TSS to switch
between the userspace and kernel stack contexts. Furthermore a basic
Syscall dispatching class was added on which Syscalls can be registered.
If a thread issues interrupt 0x80 with an id in eax, the corresponding
Syscall handler is executed with the arguments passed through ebx, ecx,
edx, esi and edi. This is the same API as Linux uses in x86.