Add class structur for the file system architecture
The filesystem architecture will be object-oriented in design. The base
File class defines the interface every type of file will have, and
almost everything will inherit this class. For example file systems will
be a File that represents where it is mounted and the files under the
file system will reference the file system.
For testing the Serial interface was changed to inherit from the
CharacterDevice class, as such it now only has one write and read
function in accordance with the File interface.