Started AHCI driver implementation
This driver is very barebones for now. It sucessfully enumerates the PCI
bus and all connected AHCI controllers. It initializes the implemented ports
on the controllers and gets the ATA IDENTIFY data from connected
devices. This now needs more work to be able to queue reads and writes,
actually dynamically crete PRDT lists, furthermore it needs locking,
suspending threads and an interface to device files.