Started VFS and Ext2 implementation
This VFS is currently very barebones and only supports rudimentary
device file actions and reading of Ext2 partitions and files.
Everything works pretty much synchronous at the moment with the
exception of blocking on I/O. But the I/O Operations themselves are not
being queued/packed together to improve disk throughput.

Furthermore there were various fixes and additions to the LibK.