Reworked kernel logging system
It now implements a design less prone to deadlocks utilizing a queue
with a shorter locked section utilizing a spinlock instead of a mutex.

Written messages are no longer immediatly printed, instead a kernel
thread reads the message queue and prints the incoming messages.

This design should make logging possible in all places. For example in
the scheduler where logging previously led to a deadlock.