Started work on the Userland
Created a project for the OwOS libc with a number of stubbed functions.
The stubbed functions make it possible to build a hosted GCC Cross
Compiler for custom i686-owos and x86_64-owos targets.

Everything in the Userland directory is built with this new toolchain
using the CMakeToolchain.txt file.

There is still work to be done on the CMake files as the libc files do
not get copied before anything else gets built which leads to link
errors.