Implemented printf for easier debugging of the linker
The printf implementation was taken from an older project of mine and
works well for all types except floating-point numbers. Still it has to
be refactored as it was not meant to be fully spec compliant and has no
error detection for failed system calls.

As the implementation of printf and ctype.h introduced some relocations
for the linker, it now performs this type of relocation on startup on
itself.